[XLA] HLO parser: fill abbreviated constant with deterministic random values.
The problem that this cl wants to solve:
Previously, if a user gives us a problematic hlo graph that OOMs, often in time we can't reproduce their issue by running the graph. This is because the abbreviated constant ("constant({...})")got interpreted as all 0, and then got optimized away into a scalar constant and broadcast (and potentially being CSE'ed).

This CL changes fills those constants with deterministic semi garbage values (byte level iota), which makes reproducing the issue easier.

PiperOrigin-RevId: 257691180