Add control-to-executor dialect conversion pass
This pass convert an MLIR representation of TensorFlow graph from the TensorFlow Control
Dialect (_tf) to a mix of tf_executor and tf dialects.
This is intended for managing the transition only, the TensorFlow Control dialect is
ultimately intended to be removed after the GraphDef importer is updated to target
directly the tf_executor dialect.

PiperOrigin-RevId: 257754326