Add TF2XLA registration for Einsum op, so that it uses XlaEinsum implementation.
This is needed because we want to switch to the fused EinsumOp in TF Classic when from the python implementation of tf.einsum. After this, the registration would enable compilation of the new TF Graphs in XLA (after cr/246570118 and compatibility window).

Note: It will still emit XlaEinsum and not Einsum when enclosing context is TPU.
PiperOrigin-RevId: 257884344