[XLA] Teach HloInstruction::OperandElementUse about kGather and fix its handling of kBitcast.
Previously OperandElementUse did not know about kGather, so it incorrectly said
gather reused both of its operands.  In fact it does a plain elementwise use of
operand 0 and does a permuting use of its operand 1.

In addition, OperandElementUse previously said that bitcast was a permuting
use.  This isn't what we want: If a fusion uses an input twice, once directly,
and once via a bitcast, that should not count as "reuse", because we're not
permuting anything, we're still going to access the elements in the same order.

PiperOrigin-RevId: 258476532