Gets base_layer_test working with run_distributed=True. Also disables autograph in the tf.function() wrapping the v2 loop.
Note that this is not being done because of any explicit bug or test failure atm, but as a conservative measure because we already autograph custom layers implemented by users. We will explore re-enabling this in the future and will use that as an opportunity to simplify built in layers, and the __call__ logic.

PiperOrigin-RevId: 258898180