Raise ValueError if an integer is passed to the training APIs.
Currently an AttributeError will be raised when x.shape is invoked in this function. In the single execution path we raise a ValueError for this. With this fix the mismatch of error types between single execution path and otherwise will be fixed. In future however, we will need to raise specific errors from Data Adapter instead of 'Failed to find data adapter that can handle ...'

PiperOrigin-RevId: 259438650