Make TensorList objects Refcounted.
This drastically reduces the amount of refcounting of individual tensors inside
TensorList when a TensorList variant is copied to a Variable or
MutableDenseHashTable (and back).  Same for operations like tf.stack that
operate on Variant tensors and perform Variant copies implicitly.

While this change adds a level of indirection into the TensorList object
by adding a heap-allocated RefCounted object to contain the vector,
it also reduces the size of the TensorList below the tf::Variant inlining
threshold.  This in turn removes a level of heap indirection and
should cancel out any performance regressions for existing
TensorList operations and small-size lists.

PiperOrigin-RevId: 259464769