Preserve element shape across TensorArray component serde
This change addresses the problem raised in #30685, where passing a TensorArray out of a tf.reduce loop would cause it to lose its inferred shape.

The issue was that when restoring the TensorArray with _from_components, we would set the _element_shape of the TensorArray wrapper class, but this field is never used. We need to set the _element_shape of the wrapped TenseorArray implementation, either _GraphTensorArray, _GraphTensorArrayV2, or _EagerTensorArray.

PiperOrigin-RevId: 259626673